requirements.txt
Flask>=3.0,<4.0
psycopg[binary]>=3.2,<4.0
gunicorn>=23.0,<24.0
python-dotenv>=1.0,<2.0
